Create README.md
## Recipo - get recipes based on ingredients you have!  
  



### Welcome to the latest recipes resource!  


1. Project Description

Recipo is an AspNetCore Web Api app, which will give you access to a multitude of recipes, based on what ingredients you select.
You have the option to use one of our recipes just with the ingredints you have in your fridge or pantry, or you could check the closest supermarket and go buy the missing ingredient or just order it by a home delivery company.
Create an account and subscribe to our app and you have the option to share your own recipes with the community.

2. Why we chose ASP.NET Core

We chose ASP.NET Core because it can run on multiple platforms, including Windows, macOS, and Linux, which makes it a flexible and scalable option for building web APIs.  Also, ASP.NET Core is designed to be fast and lightweight, which means that it can handle high traffic loads and respond quickly to client requests.
ASP.NET Core provides built-in security features, such as data validation, authorization, and authentication, which can help protect your web API from common security threats. One important factor in this decision was that it works well with popular front-end frameworks and tools, such as React, which we wanted to use and learn.

3. Challenges

One of the biggest challenges that I've faced was working with Google Cloud Platform when implementing the feature for finding the nearest supermarket. This is going to be a thing for the future to come back to and make it 100% functional.

4. How to Install and Run the Project.

Create a new project, in the IDE of your choice, and choose the template for ASP.NET Core Web Api.
Then you can clone this repository in to your project.

Then you will have to install all the following packages:
- Microsoft.AspNetCore
- Microsoft.AspNet.WebApi
- Microsoft.AspNetCore.Authentication
- Microsoft.AspNetCore.Identity
- Microsoft.AspNetCore.Identity.EntityFrameworkCore
- Microsoft.EntityFrameworkCore
- Microsoft.EntityFrameworkCore.Design
- Microsoft.EntityFrameworkCore.SqlServer
- Microsoft.EntityFrameworkCore.Tools
- Microsoft.Extensions.Identity.Core
- Microsoft.IdentityModel.Tokens
- Newtonsoft.Json
- Grpc.AspNetCore.Server
- Grpc.Core.Api
- CData.Stripe.EntityFrameworkCore
- Stripe
- Stripe.Api
- Stripe.Net
- Swashbuckle.AspNetCore


5. Credits for this project:

 - mGuy09: https://github.com/mGuy09
- me: https://github.com/MadalinaDumitrascu




  
  

<br/>  


## Rapidfire  
<table><tr><td valign="top" width="50%">

<a href="https://github.com/MadalinaDumitrascu" target="_blank">
<img src=https://img.shields.io/badge/github-%2324292e.svg?&style=for-the-badge&logo=github&logoColor=white alt=github style="margin-bottom: 5px;" />
</a>
<a href="https://linkedin.com/in/https://www.linkedin.com/in/madalina-dumitrascu-589233a0/" target="_blank">
<img src=https://img.shields.io/badge/linkedin-%231E77B5.svg?&style=for-the-badge&logo=linkedin&logoColor=white alt=linkedin style="margin-bottom: 5px;" />
</a>  


</td><td valign="top" width="50%">



</td></tr></table>  

<br/>  


## Languages and Tools  
<div align="center">  
<a href="https://reactjs.org/" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/react-original-wordmark.svg" alt="React" height="25" /></a>  
<a href="https://www.javascript.com/" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/javascript-original.svg" alt="JavaScript" height="25" /></a>  
<a href="https://github.com/" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/git-scm-icon.svg" alt="Git" height="25" /></a>  
<a href="https://docs.microsoft.com/en-us/dotnet/csharp/" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/csharp-original.svg" alt="C#" height="25" /></a>  
<a href="https://dotnet.microsoft.com/download/dotnet-framework" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/dot-net-original-wordmark.svg" alt=".NET" height="25" /></a>  
<a href="https://dotnet.microsoft.com/download" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/dotnetcore.png" alt=".Net Core" height="25" /></a>  
<a href="https://www.tailwindcss.com/" target="_blank"><img style="margin: 10px" src="https://profilinator.rishav.dev/skills-assets/tailwindcss.svg" alt="Tailwind CSS" height="25" /></a>  
</div>  

<br/>  


## Github Stats  
<table><tr><td valign="top" width="50%">

<img src="https://github-readme-stats.vercel.app/api?username=MadalinaDumitrascu&show_icons=true&count_private=true&hide_border=true" align="left" style="width: 100%" />

</td><td valign="top" width="50%">

<img src="https://github-readme-stats.vercel.app/api/top-langs/?username=MadalinaDumitrascu&hide_border=true&layout=compact" align="left" style="width: 100%" />

</td></tr></table>  

<br/>  

  

<br/>  

![Profile views counter](https://komarev.com/ghpvc/?username=mGuy09&&style=flat-square)  
  

<br/>  


<br />

----
<div align="center">Generated using <a href="https://profilinator.rishav.dev/" target="_blank">Github Profilinator</a></div>